requirements.txt
handcalcs
numpy
pandas
matplotlib
forallpeople
streamlit_drawable_canvas
pillow